// src/app/api-cinekeep/model/listCoverResponse.ts
/**
 * OpenAPI definition
 *
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */


export interface ListCoverResponse { 
    tmdbId?: number;
    mediaType?: ListCoverResponse.MediaTypeEnum;
    posterPath?: string;
    backdropPath?: string;
    selected?: boolean;
}
export namespace ListCoverResponse {
    export const MediaTypeEnum = {
        Movie: 'movie',
        Tv: 'tv'
    } as const;
    export type MediaTypeEnum = typeof MediaTypeEnum[keyof typeof MediaTypeEnum];
}